Conversion warnings fix
Fixes using ROBOPTIM_DEBUG_ONLY macro

+++ include/roboptim/trajectory/freeze.hxx
@@ -47,13 +47,13 @@ namespace trajectory
 	assert (it->first < problem_.function ().inputSize ());
 
 	Function::interval_t& interval =
-	  this->problem_.argumentBounds()[it->first];
+	  this->problem_.argumentBounds()[static_cast<std::size_t>(it->first)];
 
 	Function::value_type min =
 	  std::max (Function::getLowerBound (interval), it->second);
 	Function::value_type max =
 	  std::min (Function::getUpperBound (interval), it->second);
-	this->problem_.argumentBounds()[it->first] = Function::makeInterval (min, max);
+	this->problem_.argumentBounds()[static_cast<std::size_t>(it->first)] = Function::makeInterval (min, max);
       }
   }
